requirements.txt
absl-py==1.3.0
aioice
aiortc==1.9.0
altair==4.2.0
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
av==12.3.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click
commonmark==0.9.1
contourpy==1.0.6
cryptography
cycler==0.11.0
decorator
dnspython==2.2.1
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
grpcio
huggingface-hub
idna==3.4
importlib-metadata==5.2.0
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib
matplotlib-inline
netifaces==0.11.0
numpy
oauthlib==3.2.2
opencv-python-headless; platform_machine == 'x86_64'
packaging==22.0
pandas
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.13.0
pylibsrtp
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.0.0
rsa==4.9
scipy
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.22.0
streamlit-webrtc
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch; platform_machine == 'x86_64'
torchvision; platform_machine == 'x86_64'
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wcwidth==0.2.5
Werkzeug==2.2.2
yolov7detect==0.3.1
zipp
streamlit-drawable-canvas==0.8
svgpathtools
svgwrite
shapely
yolov7face @ git+https://github.com/securade/yolov7face
# vino and its dependencies
openvino-dev; platform_machine == 'x86_64'
openvino-telemetry; platform_machine == 'x86_64'
nncf; platform_machine == 'x86_64'
pytapo
python-telegram-bot
schedule
dash
dash_bootstrap_components
st-click-detector